+++ frontend/src/App.jsx
@@ -2,15 +2,15 @@ import { useState } from "react";
 import LandingPage from "./pages/LandingPage";
 import HomePage from "./pages/HomePage";
 import SignUp from "./pages/SignUpPage";
-import Login from "./pages/LoginPage"
+import Login from "./pages/LoginPage";
 function App() {
   return (
     <>
       {/* <LandingPage/> */}
       {/* <HomePage /> */}
       {/* <HomePage/> */}
-      {/* <SignUp /> */}
-      <Login/>
+      <SignUp />
+      {/* <Login/> */}
     </>
   );
 }
